load G2 and Capterra only after Cookie consent
