+++ src/components/HelmetIndexLayout.tsx
@@ -35,36 +35,6 @@ export const HelmetIndexLayout = ({
     <link rel="alternate" href={`${siteUrl}/fr${pathname}`} hrefLang="fr" />
     <link rel="canonical" href={`${siteUrl}${langPrefix(lang)}${pathname}`} />
     <script src="https://www.googleoptimize.com/optimize.js?id=OPT-PHR22QK" />
-    <script>
-      {`
-                (function (c, p, d, u, id, i) {
-                  id = ''; // Optional Custom ID for user in your system
-                  u = 'https://tracking.g2crowd.com/attribution_tracking/conversions/' + c + '.js?p=' + encodeURI(p) + '&e=' + id;
-                  i = document.createElement('script');
-                  i.type = 'application/javascript';
-                  i.async = true;
-                  i.src = u;
-                  d.getElementsByTagName('head')[0].appendChild(i);
-                }("3729", document.location.href, document));
-              `}
-    </script>
-    <script>
-      {`
-                (function() {
-                  var capterra_vkey = 'a12f04cb89eededafd612ba40d2d149b',
-                  capterra_vid = '2120646',
-                  capterra_prefix = (('https:' == document.location.protocol)
-                    ? 'https://ct.capterra.com' : 'http://ct.capterra.com');
-                  var ct = document.createElement('script');
-                  ct.type = 'text/javascript';
-                  ct.async = true;
-                  ct.src = capterra_prefix + '/capterra_tracker.js?vid='
-                    + capterra_vid + '&vkey=' + capterra_vkey;
-                  var s = document.getElementsByTagName('script')[0];
-                  s.parentNode.insertBefore(ct, s);
-                })();
-              `}
-    </script>
 
     {/* Disable AOS for Google */}
     <noscript>
