+++ src/helpers/index.ts
@@ -1,5 +1,4 @@
 export * from './constants';
 export * from './dynamicTranslations';
 export * from './investorFeatures';
-export { loadSegment } from './loadSegment';
 export * from './utilities';
